+++ packages/components/shared/src/mixins/hint.spec.ts
@@ -22,36 +22,50 @@ describe('HintMixin', () => {
     beforeEach(async () => {
       el = await fixture(html`<test-hint></test-hint>`);
     });
-  
+
     it('should have a hint slot', () => {
       const slot = el.renderRoot.querySelector('slot');
-  
+
       expect(slot).to.not.be.null;
       expect(slot).to.have.attribute('name', 'hint');
     });
-  
+
+    it('should have a hint slot of size medium by default', () => {
+      const slot = el.renderRoot.querySelector('slot');
+
+      expect(slot).to.have.attribute('hintSize', 'md');
+    });
+
+    it('should have a hint slot of small size when set', () => {
+      const slot = el.renderRoot.querySelector('slot');
+      slot?.setAttribute('hintsize', 'sm');
+
+      expect(slot).to.have.attribute('hintSize', 'sm');
+    });
+
     it('should render the hint text', async () => {
       el.hint = 'Hello world';
       await el.updateComplete;
-  
+
       const div = el.querySelector('div');
+
       expect(div).to.have.attribute('slot', 'hint');
       expect(div).to.have.text('Hello world');
     });
-  
+
     it('should link the input to the hint', async () => {
       el.hint = 'Hello world';
       await el.updateComplete;
-  
+
       const input = el.querySelector('input'),
         div = el.querySelector('div');
-  
+
       expect(div?.id).to.match(/sl-hint-\d+/);
       expect(input).to.have.attribute('aria-describedby', div?.id);
 
       el.hint = '';
       await el.updateComplete;
-  
+
       expect(input).not.to.have.attribute('aria-describedby');
     });
   });
@@ -60,11 +74,11 @@ describe('HintMixin', () => {
     beforeEach(async () => {
       el = await fixture(html`<test-hint><p slot="hint">Hello world!</p></test-hint>`);
     });
-  
+
     it('should link the hint to the input', async () => {
       const input = el.querySelector('input'),
         p = el.querySelector('p');
-  
+
       expect(p?.id).to.match(/sl-hint-\d+/);
       expect(input).to.have.attribute('aria-describedby', p?.id);
     });
@@ -80,29 +94,48 @@ describe('HintMixin', () => {
 
     it('should update the slotted content automatically', async () => {
       el.querySelector('p')?.remove();
-      
+
       const newHint = document.createElement('div');
       newHint.slot = 'hint';
       newHint.innerText = 'Foo bar';
       el.append(newHint);
-      
+
       await el.updateComplete;
-      
+
       const div = el.querySelector('div');
       expect(div).to.have.attribute('slot', 'hint');
       expect(div).to.have.text('Foo bar');
       expect(div?.id).to.match(/sl-hint-\d+/);
-      
+
       const input = el.querySelector('input');
       expect(input).to.have.attribute('aria-describedby', div?.id);
     });
-    
-    it('should remove the aria-describedby attribute when the hint is removed', async ()  => {
-      el.querySelector('p')?.remove();
-      await el.updateComplete;
+  });
 
+  describe('disabled', () => {
+    let el: TestHint;
+
+    beforeEach(async () => {
+      el = await fixture(html`<test-hint></test-hint>`);
+    });
+
+    it('should not be disabled by default', () => {
+      expect(el).not.to.have.attribute('disabled');
+      expect(el).not.to.match(':disabled');
+    });
+
+    it('should be disabled when the input is disabled', async () => {
       const input = el.querySelector('input');
-      expect(input).not.to.have.attribute('aria-describedby');
+
+      if (input) {
+        input.disabled = true;
+      }
+
+      el.hint = 'This is a hint';
+      await el.updateComplete;
+      const hintSlot = el.shadowRoot?.querySelector('slot[name="hint"]');
+
+      expect(hintSlot?.hasAttribute('disabled')).to.be.true;
     });
   });
 });