+++ packages/themes/bingel-int/icons.ts
@@ -1,15 +1,21 @@
 export const icons = {
   check: {
-    value: 'chevron-down',
+    value: 'check',
     type: 'Icons',
     description: 'icon.core.check',
-    svg: '<svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="M239 401c9.4 9.4 24.6 9.4 33.9 0L465 209c9.4-9.4 9.4-24.6 0-33.9s-24.6-9.4-33.9 0l-175 175L81 175c-9.4-9.4-24.6-9.4-33.9 0s-9.4 24.6 0 33.9L239 401z" fill="var(--fill-default)"></path></svg>'
+    svg: '<svg viewBox="0 0 448 512" xmlns="http://www.w3.org/2000/svg"><path d="M441 103c9.4 9.4 9.4 24.6 0 33.9L177 401c-9.4 9.4-24.6 9.4-33.9 0L7 265c-9.4-9.4-9.4-24.6 0-33.9s24.6-9.4 33.9 0l119 119L407 103c9.4-9.4 24.6-9.4 33.9 0z" fill="var(--fill-default)"></path></svg>'
   },
   'chevron-down': {
     value: 'chevron-down',
     type: 'Icons',
     description: 'icon.core.chevron-down',
     svg: '<svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="M239 401c9.4 9.4 24.6 9.4 33.9 0L465 209c9.4-9.4 9.4-24.6 0-33.9s-24.6-9.4-33.9 0l-175 175L81 175c-9.4-9.4-24.6-9.4-33.9 0s-9.4 24.6 0 33.9L239 401z" fill="var(--fill-default)"></path></svg>'
+  },
+  'circle-check-solid': {
+    value: 'circle-check',
+    type: 'Icons',
+    description: 'icon.core.circle-check-solid',
+    svg: '<svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM369 209L241 337c-9.4 9.4-24.6 9.4-33.9 0l-64-64c-9.4-9.4-9.4-24.6 0-33.9s24.6-9.4 33.9 0l47 47L335 175c9.4-9.4 24.6-9.4 33.9 0s9.4 24.6 0 33.9z" fill="var(--fill-default)"></path></svg>'
   },
   'face-smile': {
     value: 'face-smile',
@@ -52,5 +58,11 @@ export const icons = {
   },
   'sort-up': {
     svg: '<svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg"><g><path d="M8.688 1.313a.964.964 0 0 0-1.407 0l-4 4a1.01 1.01 0 0 0-.218 1.093c.156.375.53.625.937.625h8c.375 0 .75-.25.906-.625a1.01 1.01 0 0 0-.219-1.093l-4-4ZM8 13.313 5.188 10.5h5.593L8 13.313Zm-.719 1.406a.964.964 0 0 0 1.407 0l4-4c.28-.281.374-.719.218-1.094A.981.981 0 0 0 12 9H4c-.406 0-.781.25-.937.625a1.01 1.01 0 0 0 .218 1.094l4 4Z"/></g></svg>'
+  },
+  'triangle-exclamation-solid': {
+    value: 'triangle-exclamation',
+    type: 'Icons',
+    description: 'icon.core.triangle-exclamation-solid',
+    svg: '<svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="M256 32c14.2 0 27.3 7.5 34.5 19.8l216 368c7.3 12.4 7.3 27.7 .2 40.1S486.3 480 472 480H40c-14.3 0-27.6-7.7-34.7-20.1s-7-27.8 .2-40.1l216-368C228.7 39.5 241.8 32 256 32zm0 128c-13.3 0-24 10.7-24 24V296c0 13.3 10.7 24 24 24s24-10.7 24-24V184c0-13.3-10.7-24-24-24zm32 224a32 32 0 1 0 -64 0 32 32 0 1 0 64 0z" fill="var(--fill-default)"></path></svg>'
   }
 };